requirements.txt
azure-cosmos

redis

random-address

Faker